Changes according to PR feedback
